Corrected the diff's found by Volkmar Felsch's manual testing scripts. These were:
- more or less whitespace than there is in the real output,
- additional `Info' messages, because the info level wasn't reset after
  some examples (all of them were marked by <Log> ... </Log>),
  or in one instance because in the meantime the method selection chooses
  another method which doesn't produce `Info' output,
- comments in output lines,
- results from random processes, and
- one undefined variable.

There is also an example which should explicitly show what isn't possible
(product, etc. of an rcwa mapping and a permutation), and which is marked
by <Log> ... </Log> to denote that it shouldn't be tested.
The testing scripts test these examples also, but anyway, I left this
example unchanged. SK