Rcwa permutations act in a natural way on periodic lists by permuting their entries.
A corresponding action of non-bijective mappings can be defined as well --
then, entries moved to the same position are added.

In the former case (bijective mapping) the list entries can in principle
be any objects, and in the latter case it is required that list entries
can be added.

Related methods are implemented in a new file perlist.gi.
As periodic lists are implemented in the FR package,
that file is read only if FR is available. SK