Clean up category implications, fix warnings in GAP 4.11
In GAP 4.11, method installations are more closely checked for ambiguities.
This means that RCWA triggers at least one warning when loaded:

    method installed for PreImagesSet matches more than one declaration

To fix this and related things, we remove the PreImagesSet here (at
least in GAP >= 4.10; in GAP <= 4.9 we still needed); we add the
implication IsResidueClass => IsRing, and also refine the category
IsRcwaMapping to be a subset of IsMapping; moreover; several other
categories, like IsRcwaMappingOfZ, are changed to subset IsRcwaMapping
(making several InstallTrueMethod calls redundant).